refactor : LineAcceptanceTest 설정을 위한 코드를 모아서 부모 클래스로 분리
